feat: Signed BPS encoding
It is now possible to request a signed BPS encoding from the Scenes2d
Dataset. This only works for the scalar bps encoding type. In a signed
BPS encoding, basis points lying inside of an object will report a
negative value representing the distance to the nearest empty point in
the scene occupancy grid.